1.支持导出表格（没有样式） 2.支持段落，图片，表格，混合导出

+++ src/main/java/com/zjl/pdfconvert/parser/image/ImageExtractor.java
@@ -97,8 +97,8 @@ public class ImageExtractor extends PDFStreamEngine {
                 imageFact.setDisplayWidth(ctmNew.getScalingFactorX());
                 imageFact.setDisplayHeight(ctmNew.getScalingFactorY());
                 Style style = new Style();
-                style.setX(ctmNew.getTranslateX());
-                style.setY(ctmNew.getTranslateY());
+                style.setX((int)ctmNew.getTranslateX());
+                style.setY((int)ctmNew.getTranslateY());
                 imageFact.setStyle(style);
                 this.images.add(imageFact);
